showing the teaching content at teacher page and give the option of change status and delete Refs #9858

+++ removeArticle.php
@@ -1,13 +1,24 @@
 <?php
 // get the id from url
    	$aid=$_GET['aid'];
+   	$cid=$_GET['cid'];
 
 // include mysql connection file
    include 'mysql.php';
-
-// Retriving the value from article table
+   if( isset($aid))
+   {
+    // Delete the value from create_article table
 	$sql = "DELETE  FROM create_article  where aid='$aid' ";
 	$result = mysql_query( $sql,$conn);
     mysql_close($conn);
     header("location:teacher.php");
+   }
+   if( isset($cid))
+   {
+   	// Delete the value from create_content table
+	$sql = "DELETE  FROM create_content  where cid='$cid' ";
+	$result = mysql_query( $sql,$conn);
+    mysql_close($conn);
+    header("location:teacher.php");
+   }
 ?>
